change in widget realtime with datebase...[SpeedArea]
