+++ package.json
@@ -3,24 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "antd": "^4.6.5",
+    "@testing-library/jest-dom": "^5.11.6",
+    "@testing-library/react": "^11.1.2",
+    "@testing-library/user-event": "^12.2.2",
+    "antd": "^4.8.3",
     "bootstrap": "^4.5.3",
     "dotenv": "^8.2.0",
-    "firebase": "^7.24.0",
+    "firebase": "^8.0.2",
     "font-awesome": "^4.7.0",
     "fusioncharts": "^3.15.3",
     "mdbreact": "^4.27.0",
-    "react": "^16.13.1",
+    "moment": "^2.29.1",
+    "react": "^17.0.1",
     "react-bootstrap": "^1.4.0",
     "react-bootstrap-table-next": "^4.0.3",
     "react-bootstrap-table2-paginator": "^2.1.2",
     "react-dom": "^16.13.1",
     "react-fusioncharts": "^3.1.2",
+    "react-moment": "^1.0.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
+    "react-scripts": "^3.4.3",
     "react-spinners": "^0.9.0"
   },
   "scripts": {
