+++ src/component/Logs/SpeedLog.js
@@ -3,16 +3,27 @@ import { db } from "../firebase/fireConfig";
 import BootstrapTable from "react-bootstrap-table-next";
 import paginationFactory from "react-bootstrap-table2-paginator";
 import * as ReactBootstrap from "react-bootstrap";
+import Widgets from "fusioncharts/fusioncharts.widgets";
+import FusionTheme from "fusioncharts/themes/fusioncharts.theme.fusion";
+import ReactFC from "react-fusioncharts";
+import FusionCharts from "fusioncharts";
+import moment from "moment";
 
 function SpeedLog() {
+  // define chart props
+  ReactFC.fcRoot(FusionCharts, Widgets, FusionTheme);
+
   const [speedData, setSpeedData] = useState([]);
   const [loading, setLoading] = useState(false);
+
+  var lastItem;
+  var lastSpeed = 0;
 
   useEffect(() => {
     db.collection("vehicle")
       .doc("speed_sensor")
       .collection("speed")
-      .orderBy("timestamp", "asc")
+      .orderBy("id", "asc")
       .get()
       .then((snapshot) => {
         const speed_value = [];
@@ -24,6 +35,61 @@ function SpeedLog() {
       })
       .catch((error) => console.log(error));
   }, []);
+
+  lastItem = speedData[speedData.length - 1];
+  if (lastItem !== undefined) {
+    const obj = Object.entries(lastItem);
+    obj.forEach(([key, value]) => {
+      if (key === "speed") {
+        lastSpeed = value;
+      }
+    });
+  }
+
+  // config speed guage
+  const speedGuageConfigs = {
+    type: "angulargauge",
+    width: 500,
+    height: 400,
+    dataFormat: "json",
+    dataSource: {
+      chart: {
+        caption: "Vehicle Speed In [KMPH]",
+        lowerLimit: "0",
+        upperLimit: "100",
+        showValue: "1",
+        numberSuffix: "%",
+        theme: "fusion",
+        showToolTip: "0",
+      },
+      colorRange: {
+        color: [
+          {
+            minValue: "0",
+            maxValue: "50",
+            code: "#62B58F",
+          },
+          {
+            minValue: "50",
+            maxValue: "75",
+            code: "#FFC533",
+          },
+          {
+            minValue: "75",
+            maxValue: "100",
+            code: "#F2726F",
+          },
+        ],
+      },
+      dials: {
+        dial: [
+          {
+            value: lastSpeed,
+          },
+        ],
+      },
+    },
+  };
 
   const columns = [
     {
@@ -38,6 +104,7 @@ function SpeedLog() {
 
   return (
     <div>
+      <ReactFC {...speedGuageConfigs} />
       {loading ? (
         <BootstrapTable
           keyField="timestamp"
